requirements.txt
feedparser==6.0.12
requests==2.32.5
pyyaml
datetime
logging
pathlib
feedparser
apprise